debugfs: Always compile core debugfs driver for Android kernels
Unfortunately, Android userspace is very dependent on debugfs for several
unrelated things; however, it definitely doesn't require *everything*
that is included in the kernel when CONFIG_DEBUG_FS is enabled.

Therefore, in order to be able to selectively whitelist drivers that
Android needs debugfs for (by adding a #define CONFIG_DEBUG_FS to the top
of every .c file that's desired to be whitelisted), always compile the
core debugfs drivers even when CONFIG_DEBUG_FS is disabled so that
debugfs can still be used where it's necessary.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: sohamxda7 <sensoham135@gmail.com>
Signed-off-by: UsiFX <xprjkts@gmail.com>